% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_images.R
\name{extract_images}
\alias{extract_images}
\title{Extract images from Sentinel-2 archice}
\usage{
extract_images(safe_dir, id_tiles, labelled_points, .radius = 640)
}
\arguments{
\item{safe_dir}{\code{character}, path to the SAFE archives}

\item{id_tiles}{\code{character}, list of S2 tiles}

\item{labelled_points}{\code{sf}, points with a \code{label}}

\item{.radius}{\code{numeric}, half-dimension of the cropped image in map units}
}
\value{
a \code{data.frame} with information about the SiteID and file paths
}
\description{
Extract images from Sentinel-2 archice
}
