% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat2utm.R
\name{lonlat2utm}
\alias{lonlat2utm}
\title{Latlon to UTM zone}
\usage{
lonlat2utm(lonlat)
}
\arguments{
\item{lonlat}{a vector with longitude and latitude}
}
\description{
From https://bookdown.org/robinlovelace/geocompr/reproj-geo-data.html
}
