% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_geojson_polygon.R
\name{point_to_geojson_polygon}
\alias{point_to_geojson_polygon}
\title{Transform a \code{sf} point into a \code{geojson} polygon}
\usage{
point_to_geojson_polygon(point, radius = 500)
}
\arguments{
\item{point}{a \code{sf} object}

\item{radius}{\code{numeric} the radius of the spatial buffer from \code{raster::buffer()} in meters}
}
\value{
a \code{geojson} \code{Polygon}
}
\description{
Transform a \code{sf} point into a \code{geojson} polygon
}
