% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_polygon.R
\name{point_to_polygon}
\alias{point_to_polygon}
\title{Transform a \code{sf} point into a \code{geojson} polygon}
\usage{
point_to_polygon(point, radius = 5, default_crs = 4326, ...)
}
\arguments{
\item{point}{a \code{sf} object}

\item{radius}{\code{numeric} the radius of the spatial buffer from \code{raster::buffer()} in meters}

\item{default_crs}{\code{numeric}, default CRS}

\item{...}{pass to `sf::}
}
\value{
a \code{sf} \code{Polygon}
}
\description{
Transform a \code{sf} point into a \code{geojson} polygon
}
\keyword{geospatial-data}
