% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SSCT_data}
\alias{SSCT_data}
\title{Statewide channel type dataset}
\format{
A \code{data.frame}
}
\usage{
data(SSCT_data)
}
\description{
Statewide channel type dataset
}
\keyword{dataset}
