% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_polygon.R
\name{lonlat2utm}
\alias{lonlat2utm}
\title{Latlon to UTM zone
From https://bookdown.org/robinlovelace/geocompr/reproj-geo-data.html}
\usage{
lonlat2utm(lonlat)
}
\arguments{
\item{lonlat}{a two-element \code{vector} with latitude and longitude}
}
\description{
Latlon to UTM zone
From https://bookdown.org/robinlovelace/geocompr/reproj-geo-data.html
}
\keyword{geospatial-data}
